modification for browser widget to work
